refactor the debugger code into multiple libraries
