+++ packages/devtools_app/lib/src/debugger/flutter/codeview.dart
@@ -1,0 +1,264 @@
+// Copyright 2020 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:math' as math;
+
+import 'package:flutter/material.dart' hide Stack;
+import 'package:vm_service/vm_service.dart';
+
+import '../../flutter/flutter_widgets/linked_scroll_controller.dart';
+import '../../flutter/theme.dart';
+import 'debugger_controller.dart';
+
+class CodeView extends StatefulWidget {
+  const CodeView({
+    Key key,
+    this.script,
+    this.stack,
+    this.controller,
+    this.onSelected,
+    this.lineNumberToBreakpoint,
+  }) : super(key: key);
+
+  static const rowHeight = 24.0;
+  static const assumedCharacterWidth = 16.0;
+
+  final DebuggerController controller;
+  final Map<int, Breakpoint> lineNumberToBreakpoint;
+  final Script script;
+  final Stack stack;
+  final void Function(Script script, int line) onSelected;
+
+  @override
+  _CodeViewState createState() => _CodeViewState();
+}
+
+class _CodeViewState extends State<CodeView> {
+  List<String> lines = [];
+  ScrollController _horizontalController;
+  LinkedScrollControllerGroup verticalController;
+  ScrollController gutterController;
+  ScrollController textController;
+
+  // The paused positions in the current [widget.script] from the [widget.stack].
+  List<int> pausedPositions;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _horizontalController = ScrollController();
+    verticalController = LinkedScrollControllerGroup();
+    gutterController = verticalController.addAndGet();
+    textController = verticalController.addAndGet();
+    _updateLines();
+    _updatePausedPositions();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+
+    _horizontalController.dispose();
+    gutterController.dispose();
+    textController.dispose();
+  }
+
+  @override
+  void didUpdateWidget(CodeView oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    if (widget.script != oldWidget.script) {
+      _updateLines();
+    }
+
+    if (widget.stack != oldWidget.stack) {
+      _updatePausedPositions();
+    }
+  }
+
+  void _updateLines() {
+    setState(() {
+      lines = widget.script?.source?.split('\n') ?? [];
+    });
+  }
+
+  void _updatePausedPositions() {
+    setState(() {
+      final framesInScript = (widget.stack?.frames ?? [])
+          .where((frame) => frame.location.script.id == widget.script?.id);
+      pausedPositions = [
+        for (var frame in framesInScript)
+          widget.controller.lineNumber(widget.script, frame.location)
+      ];
+    });
+
+    if (pausedPositions.isNotEmpty) {
+      verticalController.animateTo(
+        pausedPositions.first * CodeView.rowHeight,
+        duration: shortDuration,
+        curve: defaultCurve,
+      );
+    }
+  }
+
+  void _onPressed(int line) {
+    widget.onSelected(widget.script, line);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO(https://github.com/flutter/devtools/issues/1648): Line numbers,
+    // syntax highlighting and breakpoint markers.
+    if (widget.script == null) {
+      return Center(
+        child: Text(
+          'No script selected',
+          style: Theme.of(context).textTheme.subtitle1,
+        ),
+      );
+    }
+
+    // Apply the log change-of-base formula, then add 16dp padding for every
+    // digit in the maximum number of lines.
+    final gutterWidth = lines.isEmpty
+        ? CodeView.assumedCharacterWidth
+        : CodeView.assumedCharacterWidth *
+                (math.log(lines.length) / math.ln10) +
+            CodeView.assumedCharacterWidth;
+
+    return DefaultTextStyle(
+      style: Theme.of(context)
+          .textTheme
+          .bodyText2
+          .copyWith(fontFamily: 'RobotoMono'),
+      child: Scrollbar(
+        child: Row(
+          children: [
+            SizedBox(
+              width: gutterWidth,
+              child: ListView(
+                controller: gutterController,
+                children: [
+                  // TODO(devoncarew): We need to virtualize this.
+                  for (var lineNum = 1; lineNum <= lines.length; lineNum++)
+                    GutterRow(
+                      lineNumber: lineNum,
+                      totalLines: lines.length,
+                      onPressed: () => _onPressed(lineNum),
+                      isBreakpoint:
+                          widget.lineNumberToBreakpoint.containsKey(lineNum),
+                    ),
+                ],
+                itemExtent: CodeView.rowHeight,
+              ),
+            ),
+            const SizedBox(width: denseSpacing),
+            Expanded(
+              child: SingleChildScrollView(
+                controller: _horizontalController,
+                scrollDirection: Axis.horizontal,
+                child: SizedBox(
+                  width: CodeView.assumedCharacterWidth *
+                      lines
+                          .map((s) => s.length)
+                          .reduce((a, b) => math.max(a, b)),
+                  child: ListView(
+                    controller: textController,
+                    children: [
+                      // Note: below, lineNum is the 1-based line number (the
+                      // user facing one, and the one that the VM uses);
+                      // lineNum - 1 is the index into the lines array.
+                      // TODO(devoncarew): We need to virtualize this.
+                      for (var lineNum = 1; lineNum <= lines.length; lineNum++)
+                        ScriptRow(
+                          lineContents: lines[lineNum - 1],
+                          onPressed: () => _onPressed(lineNum),
+                          isPausedHere: pausedPositions.contains(lineNum),
+                        )
+                    ],
+                    itemExtent: CodeView.rowHeight,
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class GutterRow extends StatelessWidget {
+  const GutterRow({
+    Key key,
+    @required this.lineNumber,
+    @required this.totalLines,
+    @required this.onPressed,
+    @required this.isBreakpoint,
+  }) : super(key: key);
+
+  final int lineNumber;
+  final int totalLines;
+  final VoidCallback onPressed;
+
+  // TODO(djshuckerow): Add support for multiple breakpoints in a line and
+  // different types of decorators than just breakpoints.
+  final bool isBreakpoint;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onPressed,
+      child: Container(
+        height: CodeView.rowHeight,
+        padding: const EdgeInsets.only(left: 2.0, right: 4.0),
+        alignment: Alignment.centerRight,
+        decoration: BoxDecoration(color: titleSolidBackgroundColor),
+        child: Row(
+          children: [
+            if (isBreakpoint) const Text('●'),
+            Expanded(
+              child: Text(
+                '$lineNumber',
+                textAlign: TextAlign.end,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class ScriptRow extends StatelessWidget {
+  const ScriptRow({
+    Key key,
+    @required this.lineContents,
+    @required this.onPressed,
+    @required this.isPausedHere,
+  }) : super(key: key);
+
+  final String lineContents;
+  final VoidCallback onPressed;
+  final bool isPausedHere;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onPressed,
+      child: Container(
+        alignment: Alignment.centerLeft,
+        height: CodeView.rowHeight,
+        color: isPausedHere ? Theme.of(context).selectedRowColor : null,
+        child: Text(
+          lineContents,
+          style: isPausedHere
+              ? TextStyle(color: Theme.of(context).textSelectionColor)
+              : null,
+        ),
+      ),
+    );
+  }
+}
