+++ packages/devtools_app/lib/src/debugger/flutter/common.dart
@@ -1,0 +1,12 @@
+// Copyright 2020 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+
+Widget densePadding(Widget child) {
+  return Padding(
+    padding: const EdgeInsets.all(2.0),
+    child: child,
+  );
+}
